requirement.txt
telebot
requests
bs4
telegram.constants